requirements.txt
# Core Flask and web framework
Flask==3.0.0
Flask-Login==0.6.3
Jinja2==3.1.2
Werkzeug==3.0.1

# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Backblaze B2 SDK
b2sdk==2.10.2

# Retry logic and utilities
tenacity==8.2.3
requests==2.31.0

# Configuration and environment
PyYAML==6.0.1
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23

# Additional utilities
python-dateutil==2.8.2
humanize==4.9.0
